requirements.txt
streamlit==1.29.0
openai-whisper==20231117
pydub==0.25.1
numpy==1.26.0
torch==2.1.0
torchaudio==2.1.0
ffmpeg-python==0.2.0 